Use NSBatchDeleteRequest for deleting all records.

+++ Data/models/CRUDProtocols.swift
@@ -41,20 +41,31 @@ public extension Deletable where Self: NSManagedObject {
     
     static func deleteAll(predicate: NSPredicate? = nil) {
         let context = DataController.newBackgroundContext()
-        let request = getFetchRequest()
-        
+        guard let request = getFetchRequest() as? NSFetchRequest<NSFetchRequestResult> else { return }
         request.predicate = predicate
         
         do {
-            let results = try context.fetch(request)
-            results.forEach {
-                context.delete($0)
+            // NSBatchDeleteRequest can't be used for in-memory store we use in tests.
+            // Have to delete objects one by one.
+            if AppConstants.IsRunningTest {
+                let results = try context.fetch(request) as? [NSManagedObject]
+                results?.forEach {
+                    context.delete($0)
+                }
+            } else {
+                let deleteRequest = NSBatchDeleteRequest(fetchRequest: request)
+                let result = try context.execute(deleteRequest) as? NSBatchDeleteResult
+                
+                let objectIDArray = result?.result as? [NSManagedObjectID]
+                let changes = [NSDeletedObjectsKey: objectIDArray as Any]
+                
+                NSManagedObjectContext.mergeChanges(fromRemoteContextSave: changes, into: [context])
             }
-            
-            DataController.save(context: context)
         } catch {
             log.error("Delete all error: \(error)")
         }
+        
+        DataController.save(context: context)        
     }
 }
